requirements.txt
# Core dependencies (yang diperlukan saja)
psycopg2-binary==2.9.9  # Database connectivity
pandas==2.0.3          # Data manipulation
minio==7.2.0           # MinIO client
python-dotenv==1.0.0   # Environment variables
Faker==22.0.0          # Data generation
python-dateutil==2.8.2 # Date utilities
